seperate ddl and cli logic
